+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Added
+
+- Type for active set of equality and inequality constraints
+
 ## [3.5.0] - 2023/08/16
 
 ### Added